*Massor

+++ ClashOfTheCharacters/ClashOfTheCharacters/Controllers/BattleController.cs
@@ -20,12 +20,12 @@ namespace ClashOfTheCharacters.Controllers
             var userId = User.Identity.GetUserId();
             var user = db.Users.Find(userId);
             var challenges = db.Challenges.Where(c => c.ChallengerId == userId || c.ReceiverId == userId && c.Accepted == false).ToList();
-            var battles = db.Battles.Where(b => b.Challenge.ChallengerId == userId || b.Challenge.ReceiverId == userId).ToList();
+            var activeBattles = db.Battles.Where(b => b.Challenge.ChallengerId == userId && !b.Calculated).ToList();
 
             ViewBag.UserId = userId;
             ViewBag.Stamina = user.Stamina;
             ViewBag.Challenges = challenges;
-            ViewBag.Battles = battles;
+            ViewBag.ActiveBattles = activeBattles;
 
             return View(db.Users.ToList());
         }
@@ -35,15 +35,17 @@ namespace ClashOfTheCharacters.Controllers
             var userId = User.Identity.GetUserId();
             var user = db.Users.Find(userId);
 
-            if (user.TeamMembers.Count == 0)
+            if (user.UserCreatures.Count == 0)
             {
                 return RedirectToAction("Select", "Character");
             }
 
             if (userId != id && id != null)
             {
+                var activeChallenges = db.Challenges.Count(c => c.ChallengerId == userId && c.ReceiverId == id && c.Accepted == false);
+                var activeBattles = db.Battles.Count(c => c.Challenge.ChallengerId == userId && c.Challenge.ReceiverId == id && c.Calculated == false);
 
-                if (user.Stamina >= 6 && db.Challenges.Count(c => c.ChallengerId == userId && c.ReceiverId == id && c.Accepted == false) < 2)
+                if (user.Stamina >= 6 && activeChallenges < 2 && activeBattles < 2)
                 {
                     if (user.Stamina == user.MaxStamina)
                     {
@@ -67,16 +69,19 @@ namespace ClashOfTheCharacters.Controllers
             var userId = User.Identity.GetUserId();
             var user = db.Users.Find(userId);
 
-            if (user.TeamMembers.Count == 0)
+            if (user.UserCreatures.Count == 0)
             {
                 return RedirectToAction("Select", "Character");
             }
 
-            var battle = new Battle { ChallengeId = id, StartTime = DateTime.Now.AddMinutes(2) };
+            if (!db.Battles.Any(b => b.ChallengeId == id))
+            {
+                var battle = new Battle { ChallengeId = id, StartTime = DateTime.Now.AddMinutes(2) };
 
-            db.Challenges.Find(id).Accepted = true;
-            db.Battles.Add(battle);
-            db.SaveChanges();
+                db.Challenges.Find(id).Accepted = true;
+                db.Battles.Add(battle);
+                db.SaveChanges();
+            }
 
             return RedirectToAction("Index");
         }
@@ -85,15 +90,18 @@ namespace ClashOfTheCharacters.Controllers
         {
             var userId = User.Identity.GetUserId();
 
-            var challenge = db.Challenges.Find(id);
-
-            if (challenge.Challenger.Stamina != challenge.Challenger.MaxStamina)
+            if (db.Challenges.Any(c => c.Id == id))
             {
-                challenge.Challenger.Stamina += 6;
-            }
+                var challenge = db.Challenges.Find(id);
 
-            db.Challenges.Remove(challenge);
-            db.SaveChanges();
+                if (challenge.Challenger.Stamina != challenge.Challenger.MaxStamina)
+                {
+                    challenge.Challenger.Stamina += 6;
+                }
+
+                db.Challenges.Remove(challenge);
+                db.SaveChanges();
+            }
 
             return RedirectToAction("Index");
         }
